Use readall() in TAsyncSocket to ensure all data is received
