+++ thriftpy2/contrib/aio/transport/buffered.py
@@ -2,23 +2,8 @@
 import asyncio
 from io import BytesIO
 
-from thriftpy2.transport import TTransportBase, TTransportException
-
-
-@asyncio.coroutine
-def readall(read_fn, sz):
-    buff = b''
-    have = 0
-    while have < sz:
-        chunk = yield from read_fn(sz - have)
-        have += len(chunk)
-        buff += chunk
-
-        if len(chunk) == 0:
-            raise TTransportException(TTransportException.END_OF_FILE,
-                                      "End of file reading from transport")
-
-    return buff
+from thriftpy2.transport import TTransportBase
+from ..socket import readall
 
 
 class TAsyncBufferedTransport(TTransportBase):
